backend/src/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-gis==1.0
environ==1.0
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
